+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The changelog for **SwifterSwift**. Also see the [releases](https://github.com/S
 ## Upcoming Release
 
 ### Added
+- **CLLocation**
+      - Added `placemarks(completion:)` to `CLLocation`. [#806](https://github.com/SwifterSwift/SwifterSwift/pull/806) by [trevorphillips](https://github.com/trevorphillips)
 - **Optional**:
   - Conform to `Equatable` when `Wrapped` is `RawRepresentable` and its `RawValue` is `Equatable`. [#804](https://github.com/SwifterSwift/SwifterSwift/pull/804) by [guykogus](https://github.com/guykogus)
 - **CoreLocation**: